Splitting classes in separated files
