+++ front/src/Api/iframe/ui.ts
@@ -2,39 +2,9 @@ import { isButtonClickedEvent } from '../Events/ButtonClickedEvent';
 import type { ClosePopupEvent } from '../Events/ClosePopupEvent';
 import { IframeApiContribution, sendToWorkadventure } from './IframeApiContribution';
 import { apiCallback } from "./registeredCallbacks";
+import {Popup} from "./Ui/Popup";
+import type {ButtonClickedCallback, ButtonDescriptor} from "./Ui/ButtonDescriptor";
 
-export class Popup {
-    constructor(private id: number) {
-    }
-
-    /**
-     * Closes the popup
-     */
-    public close(): void {
-        sendToWorkadventure({
-            'type': 'closePopup',
-            'data': {
-                'popupId': this.id,
-            } as ClosePopupEvent
-        });
-    }
-}
-
-type ButtonClickedCallback = (popup: Popup) => void;
-export interface ButtonDescriptor {
-    /**
-     * The label of the button
-     */
-    label: string,
-    /**
-     * The type of the button. Can be one of "normal", "primary", "success", "warning", "error", "disabled"
-     */
-    className?: "normal" | "primary" | "success" | "warning" | "error" | "disabled",
-    /**
-     * Callback called if the button is pressed
-     */
-    callback: ButtonClickedCallback,
-}
 let popupId = 0;
 const popups: Map<number, Popup> = new Map<number, Popup>();
 const popupCallbacks: Map<number, Map<number, ButtonClickedCallback>> = new Map<number, Map<number, ButtonClickedCallback>>();
@@ -50,9 +20,6 @@ interface ZonedPopupOptions {
 
 class WorkAdventureUiCommands extends IframeApiContribution<WorkAdventureUiCommands> {
 
-    readonly subObjectIdentifier = "ui"
-
-    readonly addMethodsAtRoot = true
     callbacks = [apiCallback({
         type: "buttonClickedEvent",
         typeChecker: isButtonClickedEvent,
